# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DealVault API",
    "version": "1.0.0",
    "description": "Commodity deal room management platform API. All endpoints require authentication via NextAuth session unless noted otherwise."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Development" }
  ],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      }
    },
...
  },
  "security": [{ "session": [] }],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string", "minLength": 2 },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8 },
                  "phone": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created successfully" },
          "400": { "description": "Validation error or email already exists" }
        }
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Request password reset token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": { "type": "string", "format": "email" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Reset email sent (always returns success to prevent enumeration)" }
        }
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Reset password with token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token", "password"],
                "properties": {
                  "token": { "type": "string" },
                  "password": { "type": "string", "minLength": 8 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Password reset successfully" },
          "400": { "description": "Invalid or expired token" }
        }
      }
    },
...
    "/api/deals/{id}/commission": {
      "get": {
        "tags": ["Commission"],
        "summary": "List commission ledger entries",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Array of commission entries" }
        }
      },
      "post": {
        "tags": ["Commission"],
        "summary": "Set/update commission for a party",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["partyId", "agreedPct"],
                "properties": {
                  "partyId": { "type": "string" },
                  "agreedPct": { "type": "number", "minimum": 0, "description": "Decimal (e.g. 0.005 = 0.5%)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Commission set/updated" },
          "400": { "description": "Exceeds commission pool" }
        }
      }
    },
...
    "/api/deals/{id}/export": {
      "get": {
        "tags": ["Deals"],
        "summary": "Export audit log as CSV or JSON",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "format", "in": "query", "schema": { "type": "string", "enum": ["csv", "json"], "default": "json" } }
        ],
        "responses": {
          "200": { "description": "Audit log export file" }
        }
      }
    },
...
    "/api/companies": {
      "get": {
        "tags": ["Companies"],
        "summary": "List all companies created by user",
        "responses": {
          "200": { "description": "Array of companies" }
        }
      },
      "post": {
        "tags": ["Companies"],
        "summary": "Create a company",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "minLength": 2 },
                  "registrationNumber": { "type": "string" },
                  "country": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Company created" }
        }
      }
    },
    "/api/companies/{id}": {
      "patch": {
        "tags": ["Companies"],
        "summary": "Update a company",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Company updated" } }
      },
      "delete": {
        "tags": ["Companies"],
        "summary": "Delete a company (fails if linked to deals)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Company deleted" },
          "400": { "description": "Company is linked to active deals" }
        }
      }
    },
...
  }
}